code/requirements.txt
psutil>=5.8.0
pandas>=1.1.3
torch>=1.8.0
matplotlib>=3.3.4
timm>=0.3.4
fastcore>=1.3.19
catalyst>=20.12
torchvision>=0.9.0
numpy>=1.19.2
transformers>=4.2.2
pytorch_lightning==1.1.8
seaborn>=0.11.1
Pillow>=8.1.2
scikit_learn>=0.24.1
